+++ src/badge/interfaces.ts
@@ -3,10 +3,21 @@ export const BadgeVariants = {
 	outlined: 'outlined',
 };
 
+export const BadgeColor = {
+	primary: 'primary',
+	grey: 'grey',
+};
+
+export interface BadgeColorValues {
+	hover: string;
+	base: string;
+}
+
 export interface BadgeProps {
 	leftIcon?: React.ReactElement;
 	rightIcon?: React.ReactElement;
-	text: string;
+	text: string | React.ReactElement;
 	variant?: keyof typeof BadgeVariants;
+	color?: keyof typeof BadgeColor;
 	onClick?: () => void;
 }