+++ src/index.ts
@@ -7,3 +7,5 @@ export * from './badge';
 export * from './dropdown';
 export * from './portal';
 export * from './modal-window';
+export * from './navigation';
+export * from './layout';