+++ src/input/styles.ts
@@ -26,6 +26,14 @@ export const StyledInputWrapper = styled('div')<
 	}};
 	padding: 12px;
 
+	${(props) => {
+		if (props.variant === 'large') {
+			return `min-height: 30px;`;
+		}
+
+		return ``;
+	}}
+
 	border: 2px solid ${(props) => props.theme.colors.grey[600]};
 	border-radius: 8px;
 	background-color: ${(props) => props.theme.colors.grey[100]};