+++ src/layout/layout-full-window/layout-full-window.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { StyledFullLayout } from './styles';
+
+export const LayoutFullWindow: React.FC<{
+	children: React.ReactElement | React.ReactNode;
+	gridTemplage?: string;
+	className?: string;
+}> = ({ children, gridTemplage, className }): React.ReactElement => {
+	return (
+		<StyledFullLayout className={className} gridTemplate={gridTemplage}>
+			{children}
+		</StyledFullLayout>
+	);
+};