+++ src/navigation/components/navigation-option/index.ts
@@ -1,4 +1,3 @@
-export * from './input';
+export * from './navigation-option';
 export * from './interfaces';
 export * from './styles';
-export * from './components';