+++ src/navigation/index.ts
@@ -1,4 +1,4 @@
-export * from './input';
+export * from './navigation';
 export * from './interfaces';
-export * from './styles';
 export * from './components';
+export * from './styles';