+++ src/stories/Navigation.stories.tsx
@@ -1,11 +1,13 @@
+import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 import { fn } from '@storybook/test';
-import { Dropdown } from '../dropdown';
+import { Navigation } from '../navigation';
+import { LayoutFullWindow } from '../layout';
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
-	title: 'Components/Dropdown',
-	component: Dropdown,
+	title: 'Components/Navigation',
+	component: Navigation,
 	parameters: {
 		// Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
 		layout: 'centered',
@@ -16,17 +18,10 @@ const meta = {
 	// More on argTypes: https://storybook.js.org/docs/api/argtypes
 	// Use `fn` to spy on the onClick arg, which will appear in the actions panel once invoked: https://storybook.js.org/docs/essentials/actions#action-args
 	args: {
-		options: [
-			{ title: 'test1', value: 'test1', id: '1' },
-			{ title: 'test2', value: 'test2', id: '2' },
-		],
-		onSelect: fn(),
-		value: [
-			{ title: 'test1', value: 'test1', id: '1' },
-			{ title: 'test2', value: 'test2', id: '2' },
-		],
+		onSelectOption: fn(),
+		options: [{ text: 'Home', link: '/home', selected: true }],
 	},
-} satisfies Meta<typeof Dropdown>;
+} satisfies Meta<typeof Navigation>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
@@ -34,18 +29,18 @@ type Story = StoryObj<typeof meta>;
 // More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
 export const Primary: Story = {
 	args: {
-		placeholder: 'Выберите одну из опций',
-		options: [
-			{ title: 'test1', value: 'test1', id: '1' },
-			{ title: 'test2', value: 'test2', id: '2' },
-			{ title: 'test3', value: 'test3', id: '3' },
-			{ title: 'test4', value: 'test4', id: '4' },
-		],
 		onSelectOption: fn(),
-		value: [
-			// { title: 'test1', value: 'test1', id: '1' },
-			{ title: 'test2', value: 'test2', id: '2' },
-		],
-		multiSelect: false,
+		logo: 'Logo',
+		options: [{ text: 'Home', link: '/home', selected: true }],
+	},
+	render: (args) => {
+		return (
+			<div style={{ width: '1000px' }}>
+				<LayoutFullWindow>
+					<Navigation {...args} />
+					<div>content</div>
+				</LayoutFullWindow>
+			</div>
+		);
 	},
 };